requirements.txt
django
djangorestframework
factory_boy
pytest
django-redis
django-environ
psycopg2
faker
pytest-django